+++ briar-core/src/main/java/org/briarproject/briar/sharing/SharingValidator.java
@@ -40,7 +40,7 @@ abstract class SharingValidator extends BdfMessageValidator {
 	@Override
 	protected BdfMessageContext validateMessage(Message m, Group g,
 			BdfList body) throws FormatException {
-		MessageType type = MessageType.fromValue(body.getLong(0).intValue());
+		MessageType type = MessageType.fromValue(body.getInt(0));
 		switch (type) {
 			case INVITE:
 				return validateInviteMessage(m, body);
